push: split push up into locale_file and push_source
